Refactor Before Parallel PPS change
